* Increase geometry simplification threshold from 0.01 to 0.02
* Optimize queryset retrieval in views to load only necessary fields
* Update serializers accordingly

> **Note**: The geometry simplification should be discussed with the client to decide on an agreeable threshold. 

Performance was gauged by triggering the request through the frontend and measuring the average request time and database time over 20 requests. The performance improved from **4682 ms request time** and **240ms DB time** to **190 ms request time** and **21 ms DB time**.